implemented pending screen within login component
