--- pupil-spa/src/app/login-pending/login-pending.component.html
@@ -1,14 +1,0 @@
-<app-header class="full-width"></app-header>
-<main id="content" class="js-content" role="main">
-  <div class="grid-row">
-    <div class="column-full">
-      <div class="login-description">
-        <app-svg-loading aria-hidden="true"></app-svg-loading>
-        <span class="wait-text aa-lead-size" tabindex="0">
-            <b class="font-large">Loading...</b>
-          </span>
-      </div>
-    </div>
-  </div>
-</main>
-<app-footer></app-footer>
