--- pupil-spa/src/app/login-pending/login-pending.component.scss
@@ -1,17 +1,0 @@
-.login-description {
-  margin-top: 10vh;
-  text-align: center;
-  min-height: 500px;
-}
-
-app-svg-loading {
-  display: inline-block;
-  margin-bottom: 20px;
-}
-
-.wait-text {
-  left: 5px;
-  top: 2px;
-  position: relative;
-  font-size: x-large;
-}
