+++ pupil-spa/src/app/login/login.component.ts
@@ -1,10 +1,16 @@
 import { Component, ElementRef, OnInit, AfterViewInit } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
+import { APP_CONFIG } from '../services/config/config.service';
+import { Router } from '@angular/router';
 
 import { LoginErrorService } from '../services/login-error/login-error.service';
 import { LoginErrorDiagnosticsService } from '../services/login-error-diagnostics/login-error-diagnostics.service';
+import { UserService } from '../services/user/user.service';
+import { QuestionService } from '../services/question/question.service';
+import { WarmupQuestionService } from '../services/question/warmup-question.service';
+import { RegisterInputService } from '../services/register-input/registerInput.service';
 import { CheckStatusService } from '../services/check-status/check-status.service';
 import { Login } from './login.model';
+import { PupilPrefsService } from '../services/pupil-prefs/pupil-prefs.service';
 
 @Component({
   selector: 'app-login',
@@ -12,27 +18,38 @@ import { Login } from './login.model';
   styleUrls: ['./login.component.scss']
 })
 export class LoginComponent implements OnInit, AfterViewInit {
+
+  private submitted: boolean;
   public loginModel = new Login('', '');
+  public loginPending: boolean;
   public loginSucceeded: boolean;
+  public connectionFailed: boolean;
+  public loginPendingViewMinDisplay: number;
   private errorMessage: string;
 
   constructor(
     private loginErrorService: LoginErrorService,
     private loginErrorDiagnosticsService: LoginErrorDiagnosticsService,
+    private userService: UserService,
     private router: Router,
-    private route: ActivatedRoute,
+    private questionService: QuestionService,
+    private warmupQuestionService: WarmupQuestionService,
     private elRef: ElementRef,
-    private checkStatusService: CheckStatusService
-  ) { }
+    private registerInputService: RegisterInputService,
+    private checkStatusService: CheckStatusService,
+    private pupilPrefsService: PupilPrefsService
+  ) {
+    const { loginPendingViewMinDisplay } = APP_CONFIG;
+    this.loginPendingViewMinDisplay = loginPendingViewMinDisplay;
+  }
 
   ngOnInit() {
+    this.loginPending = false;
     const hasUnfinishedCheck = this.checkStatusService.hasUnfinishedCheck();
     if (hasUnfinishedCheck) {
       this.router.navigate(['check'], { queryParams: { unfinishedCheck: true } });
     }
     this.loginErrorService.currentErrorMessage.subscribe(message => this.errorMessage = message);
-    const queryParams = this.route.snapshot.queryParams;
-    this.loginSucceeded = queryParams && queryParams.loginSucceeded && JSON.parse(queryParams.loginSucceeded);
   }
 
   ngAfterViewInit() {
@@ -53,6 +70,75 @@ export class LoginComponent implements OnInit, AfterViewInit {
    * Handler for the login form submit action
    */
   onSubmit(schoolPin, pupilPin) {
-    this.router.navigate(['sign-in-pending'], { queryParams: { schoolPin, pupilPin } });
+    const startTime = Date.now();
+    this.loginPending = true;
+    if (this.submitted === true) {
+      return;
+    }
+    this.submitted = true;
+    this.userService.login(schoolPin, pupilPin)
+      .then(
+        async () => {
+          this.loginSucceeded = true;
+          this.connectionFailed = false;
+          this.questionService.initialise();
+          this.warmupQuestionService.initialise();
+          this.registerInputService.initialise();
+
+          const config = this.questionService.getConfig();
+          this.pupilPrefsService.loadPupilPrefs();
+          await this.displayMinTime(startTime);
+          if (config.fontSize) {
+            this.router.navigate(['font-choice']);
+          } else if (config.colourContrast) {
+            this.router.navigate(['colour-choice']);
+          } else {
+            this.router.navigate(['sign-in-success']);
+          }
+        },
+        async (err) => {
+          this.submitted = false;
+          this.loginErrorService.changeMessage(err.message);
+          if (err.status === 401) {
+            await this.displayMinTime(startTime);
+            this.loginSucceeded = false;
+            this.loginPending = false;
+            this.router.navigate(['sign-in']);
+          } else {
+            await this.loginErrorDiagnosticsService.process(err);
+            this.router.navigate(['sign-in-fail']);
+          }
+        })
+      .catch(async () => {
+        await this.displayMinTime(startTime);
+        this.loginSucceeded = false;
+        this.submitted = false;
+        this.loginPending = false;
+        this.router.navigate(['sign-in']);
+      });
+  }
+
+  /**
+   * Display login pending screen for minimum duration
+   * @param {Number} startTime
+   * @returns {Promise.<void>}
+   */
+  async displayMinTime(startTime) {
+    const endTime = Date.now();
+    const duration = endTime - startTime;
+    const minDisplay = this.loginPendingViewMinDisplay;
+    if (duration < minDisplay) {
+      const displayTime = minDisplay - duration;
+      return this.sleep(displayTime);
+    }
+  }
+
+  /**
+   * Sleep function (milliseconds) to provide minimal display time for submission pending screen
+   * @param {Number} ms
+   * @returns {Promise.<void>}
+   */
+  private sleep(ms) {
+    return new Promise(resolve => setTimeout(resolve, ms));
   }
 }
